Added new presenter for periodic table and peaks
