+++ scripts/Muon/GUI/ElementalAnalysis/elemental_analysis.py
@@ -4,37 +4,24 @@
 #   NScD Oak Ridge National Laboratory, European Spallation Source,
 #   Institut Laue - Langevin & CSNS, Institute of High Energy Physics, CAS
 # SPDX - License - Identifier: GPL - 3.0 +
-from qtpy import QtWidgets
 from copy import deepcopy
+
 import matplotlib as mpl
-
-from Muon.GUI.ElementalAnalysis.PeriodicTable.periodic_table_presenter import PeriodicTablePresenter
-from Muon.GUI.ElementalAnalysis.PeriodicTable.periodic_table_view import PeriodicTableView
-from Muon.GUI.ElementalAnalysis.PeriodicTable.periodic_table_model import PeriodicTableModel
-
-from MultiPlotting.multi_plotting_widget import MultiPlotWindow
-from MultiPlotting.label import Label
-
-from Muon.GUI.ElementalAnalysis.LoadWidget.load_model import LoadModel, CoLoadModel
-from Muon.GUI.ElementalAnalysis.LoadWidget.load_utils import spectrum_index
-
-from Muon.GUI.Common.load_widget.load_view import LoadView
-from Muon.GUI.Common.load_widget.load_presenter import LoadPresenter
-
-from Muon.GUI.ElementalAnalysis.Detectors.detectors_presenter import DetectorsPresenter
-from Muon.GUI.ElementalAnalysis.Detectors.detectors_view import DetectorsView
-from Muon.GUI.ElementalAnalysis.Peaks.peaks_presenter import PeaksPresenter
-from Muon.GUI.ElementalAnalysis.Peaks.peaks_view import PeaksView
-
-from Muon.GUI.ElementalAnalysis.PeriodicTable.PeakSelector.peak_selector_presenter import PeakSelectorPresenter
-from Muon.GUI.ElementalAnalysis.PeriodicTable.PeakSelector.peak_selector_view import PeakSelectorView
-
-from Muon.GUI.ElementalAnalysis.LineSelector.LineSelectorPresenter import LineSelectorPresenter
-from Muon.GUI.ElementalAnalysis.LineSelector.LineSelectorView import LineSelectorView
-
-from Muon.GUI.Common import message_box
+from qtpy import QtWidgets
 
 import mantid.simpleapi as mantid
+from MultiPlotting.multi_plotting_widget import MultiPlotWindow
+from Muon.GUI.Common.load_widget.load_presenter import LoadPresenter
+from Muon.GUI.Common.load_widget.load_view import LoadView
+from Muon.GUI.ElementalAnalysis.Detectors.detectors_presenter import DetectorsPresenter
+from Muon.GUI.ElementalAnalysis.Detectors.detectors_view import DetectorsView
+from Muon.GUI.ElementalAnalysis.LineSelector.LineSelectorPresenter import LineSelectorPresenter
+from Muon.GUI.ElementalAnalysis.LineSelector.LineSelectorView import LineSelectorView
+from Muon.GUI.ElementalAnalysis.LoadWidget.load_model import LoadModel, CoLoadModel
+from Muon.GUI.ElementalAnalysis.LoadWidget.load_utils import spectrum_index
+from Muon.GUI.ElementalAnalysis.Peaks.peaks_view import PeaksView
+from Muon.GUI.ElementalAnalysis.PeriodicTable.periodic_table_view import PeriodicTableView
+from Muon.GUI.ElementalAnalysis.periodic_table_widget_presenter import PeriodicTableWidgetPresenter
 
 offset = 0.9
 
@@ -71,18 +58,6 @@ class ElementalAnalysisGui(QtWidgets.QMainWindow):
         super(ElementalAnalysisGui, self).__init__(parent)
         if window_flags:
             self.setWindowFlags(window_flags)
-        # set menu
-        self.menu = self.menuBar()
-        self.menu.addAction("File")
-        edit_menu = self.menu.addMenu("Edit")
-        edit_menu.addAction("Change Peak Data file", self.select_data_file)
-        self.menu.addAction("Binning")
-        self.menu.addAction("Normalise")
-
-        # periodic table stuff
-        self.ptable = PeriodicTablePresenter(PeriodicTableView(), PeriodicTableModel())
-        self.ptable.register_table_lclicked(self.table_left_clicked)
-        self.ptable.register_table_rclicked(self.table_right_clicked)
 
         # load stuff
         self.load_widget = LoadPresenter(LoadView(), LoadModel(), CoLoadModel())
@@ -94,18 +69,24 @@ class ElementalAnalysisGui(QtWidgets.QMainWindow):
             detector.on_checkbox_checked(self.add_plot)
             detector.on_checkbox_unchecked(self.del_plot)
 
-        # peaks boxes
-        self.peaks = PeaksPresenter(PeaksView())
-        self.peaks.major.setChecked(True)
-        self.peaks.major.on_checkbox_checked(self.major_peaks_changed)
-        self.peaks.major.on_checkbox_unchecked(self.major_peaks_changed)
-        self.peaks.minor.on_checkbox_checked(self.minor_peaks_changed)
-        self.peaks.minor.on_checkbox_unchecked(self.minor_peaks_changed)
-        self.peaks.gamma.on_checkbox_checked(self.gammas_changed)
-        self.peaks.gamma.on_checkbox_unchecked(self.gammas_changed)
-        self.peaks.electron.on_checkbox_checked(self.electrons_changed)
-        self.peaks.electron.on_checkbox_unchecked(self.electrons_changed)
-        self.peaks.set_deselect_elements_slot(self.deselect_elements)
+        # plotting
+        self._plot_window = None
+        self._plotting = None
+        self.num_colors = len(mpl.rcParams['axes.prop_cycle'])
+        self.used_colors = {}
+
+        # setup periodic table and peaks
+        self.ptable_view = PeriodicTableView()
+        self.peaks_view = PeaksView()
+        self.ptable = PeriodicTableWidgetPresenter(self)
+
+        # set menu
+        self.menu = self.menuBar()
+        self.menu.addAction("File")
+        edit_menu = self.menu.addMenu("Edit")
+        edit_menu.addAction("Change Peak Data file", self.ptable.select_data_file)
+        self.menu.addAction("Binning")
+        self.menu.addAction("Normalise")
 
         # Line type boxes
         self.lines = LineSelectorPresenter(LineSelectorView())
@@ -119,29 +100,37 @@ class ElementalAnalysisGui(QtWidgets.QMainWindow):
 
         # set up
         self.widget_list = QtWidgets.QVBoxLayout()
-        self.widget_list.addWidget(self.peaks.view)
+        self.widget_list.addWidget(self.peaks_view)
         self.widget_list.addWidget(self.lines.view)
         self.widget_list.addWidget(self.detectors.view)
         self.widget_list.addWidget(self.load_widget.view)
 
-        # plotting
-        self.plot_window = None
-        self.num_colors = len(mpl.rcParams['axes.prop_cycle'])
-        self.used_colors = {}
-
         # layout
         self.box = QtWidgets.QHBoxLayout()
-        self.box.addWidget(self.ptable.view)
+        self.box.addWidget(self.ptable_view)
         self.box.addLayout(self.widget_list)
 
         self.setCentralWidget(QtWidgets.QWidget(self))
         self.centralWidget().setLayout(self.box)
         self.setWindowTitle("Elemental Analysis")
 
-        self.element_widgets = {}
-        self.element_lines = {}
-        self.electron_peaks = {}
-        self._generate_element_widgets()
+    @property
+    def plot_window(self):
+        return self._plot_window
+
+    @plot_window.setter
+    def plot_window(self, plot_window):
+        self._plot_window = plot_window
+        self.ptable.plot_window = plot_window
+
+    @property
+    def plotting(self):
+        return self._plotting
+
+    @plotting.setter
+    def plotting(self, plotting):
+        self._plotting = plotting
+        self.ptable.plotting = plotting
 
     def get_color(self, element):
         """
@@ -174,93 +163,6 @@ class ElementalAnalysisGui(QtWidgets.QMainWindow):
         if self.plot_window is not None:
             self.plot_window.closeEvent(event)
         super(ElementalAnalysisGui, self).closeEvent(event)
-
-    # general functions
-
-    def _gen_label(self, name, x_value_in, element=None):
-        if element is None:
-            return
-        # check x value is a float
-        try:
-            x_value = float(x_value_in)
-        except:
-            return
-        if name not in self.element_lines[element]:
-            self.element_lines[element].append(name)
-        # make sure the names are strings and x values are numbers
-        return Label(str(name), x_value, False, offset, True, rotation=-90, protected=True)
-
-    def _plot_line(self, name, x_value_in, color, element=None):
-        label = self._gen_label(name, x_value_in, element)
-        if self.plot_window is None:
-            return
-        for subplot in self.plotting.get_subplots():
-            self._plot_line_once(subplot, x_value_in, label, color)
-
-    def _plot_line_once(self, subplot, x_value, label, color):
-        self.plotting.add_vline_and_annotate(subplot, x_value, label, color)
-
-    def _rm_line(self, name):
-        if self.plot_window is None:
-            return
-        for subplot in self.plotting.get_subplots():
-            self.plotting.rm_vline_and_annotate(subplot, name)
-
-    # setup element pop up
-    def _generate_element_widgets(self):
-        self.element_widgets = {}
-        for element in self.ptable.peak_data:
-            data = self.ptable.element_data(element)
-            widget = PeakSelectorPresenter(PeakSelectorView(data, element))
-            widget.on_finished(self._update_peak_data)
-            self.element_widgets[element] = widget
-
-    # interact with periodic table
-    def table_right_clicked(self, item):
-        self.element_widgets[item.symbol].view.show()
-
-    def table_left_clicked(self, item):
-        if self.ptable.is_selected(item.symbol):
-            self._add_element_lines(item.symbol)
-        else:
-            self._remove_element_lines(item.symbol)
-
-    def _add_element_lines(self, element, data=None):
-        if data is None:
-            data = self.element_widgets[element].get_checked()
-        if element not in self.element_lines:
-            self.element_lines[element] = []
-
-        # Select a different color, if all used then use the first
-        color = self.get_color(element)
-
-        for name, x_value in data.items():
-            try:
-                x_value = float(x_value)
-            except ValueError:
-                continue
-            full_name = gen_name(element, name)
-            if full_name not in self.element_lines[element]:
-                self.element_lines[element].append(full_name)
-            self._plot_line(full_name, x_value, color, element)
-
-    def _remove_element_lines(self, element):
-        if element not in self.element_lines:
-            return
-        names = deepcopy(self.element_lines[element])
-        for name in names:
-            try:
-                self.element_lines[element].remove(name)
-                self._rm_line(name)
-                if not self.element_lines[element]:
-                    del self.element_lines[element]
-                    if element in self.used_colors:
-                        del self.used_colors[element]
-            except:
-                continue
-
-        if element in self.element_lines and not self.element_lines[element]:
-            del self.element_lines[element]
 
     # loading
     def load_run(self, detector, run):
@@ -322,7 +224,7 @@ class ElementalAnalysisGui(QtWidgets.QMainWindow):
             self.plotting.plot(detector, ws.name(), spec_num=spectrum_index["Delayed"], color=self.GREEN)
 
         # add current selection of lines
-        for element in self.ptable.selection:
+        for element in self.ptable.ptable.selection:
             self.add_peak_data(element.symbol, detector)
 
     def _unset_detectors(self):
@@ -332,34 +234,13 @@ class ElementalAnalysisGui(QtWidgets.QMainWindow):
             self.detectors.setStateQuietly(name, False)
         self.uncheck_detectors_if_no_line_plotted()
 
-    # plotting
-    def add_peak_data(self, element, subplot, data=None):
-        # if already selected add to just new plot
-        if data is None:
-            data = self.element_widgets[element].get_checked()
-        color = self.get_color(element)
-        for name, x_value in data.items():
-            if isinstance(x_value, float):
-                full_name = gen_name(element, name)
-                label = self._gen_label(full_name, x_value, element)
-                self._plot_line_once(subplot, x_value, label, color)
-
-    def _update_peak_data(self, element):
-        if self.ptable.is_selected(element):
-            # remove all of the lines for the element
-            self._remove_element_lines(element)
-            # add the ticked lines to the plot
-            self._add_element_lines(element)
-        else:
-            self._remove_element_lines(element)
-
     def add_plot(self, checkbox):
         detector = checkbox.name
         last_run = self.load_widget.last_loaded_run()
         # not using load_last_run prevents two calls to last_loaded_run()
         if last_run is not None:
             self.load_run(detector, last_run)
-            self._update_checked_data()
+            self.ptable._update_checked_data()
 
     def del_plot(self, checkbox):
         if self.load_widget.last_loaded_run() is not None:
@@ -375,69 +256,6 @@ class ElementalAnalysisGui(QtWidgets.QMainWindow):
         if not self.plotting.get_subplots():
             self.plot_window.close()
             self.plot_window = None
-
-    # sets data file for periodic table
-    def select_data_file(self):
-        old_lines = deepcopy(list(self.element_lines.keys()))
-
-        filename = QtWidgets.QFileDialog.getOpenFileName()
-        if isinstance(filename, tuple):
-            filename = filename[0]
-        filename = str(filename)
-        if filename:
-            self.ptable.set_peak_datafile(filename)
-
-        try:
-            self._generate_element_widgets()
-        except ValueError:
-            message_box.warning(
-                'The file does not contain correctly formatted data, resetting to default data file.'
-                'See "https://docs.mantidproject.org/nightly/interfaces/'
-                'Muon%20Elemental%20Analysis.html" for more information.')
-            self.ptable.set_peak_datafile(None)
-            self._generate_element_widgets()
-
-        for element in old_lines:
-            if element in self.element_widgets.keys():
-                self.ptable.select_element(element)
-            else:
-                self._remove_element_lines(element)
-        self._update_checked_data()
-
-    def _update_checked_data(self):
-        self.major_peaks_changed(self.peaks.major)
-        self.minor_peaks_changed(self.peaks.minor)
-        self.gammas_changed(self.peaks.gamma)
-        self.electrons_changed(self.peaks.electron)
-
-    # general checked data
-    def checked_data(self, element, selection, state):
-        for checkbox in selection:
-            checkbox.setChecked(state)
-        self._update_peak_data(element)
-
-    def electrons_changed(self, electron_peaks):
-        for element, selector in self.element_widgets.items():
-            self.checked_data(element, selector.electron_checkboxes, electron_peaks.isChecked())
-
-    def gammas_changed(self, gamma_peaks):
-        for element, selector in self.element_widgets.items():
-            self.checked_data(element, selector.gamma_checkboxes, gamma_peaks.isChecked())
-
-    def major_peaks_changed(self, major_peaks):
-        for element, selector in self.element_widgets.items():
-            self.checked_data(element, selector.primary_checkboxes, major_peaks.isChecked())
-
-    def minor_peaks_changed(self, minor_peaks):
-        for element, selector in self.element_widgets.items():
-            self.checked_data(element, selector.secondary_checkboxes, minor_peaks.isChecked())
-
-    def deselect_elements(self):
-        self.peaks.disable_deselect_elements_btn()
-        for element in self.element_widgets.keys():
-            self.ptable.deselect_element(element)
-            self._remove_element_lines(element)
-        self.peaks.enable_deselect_elements_btn()
 
     def add_line_by_type(self, run, _type):
         # Ensure all detectors are enabled
